Updated favicon, .gitignore, and readme
small commit to help the favicon out and instruct users on how to use venv to organize the requirements locally in their repo without installing in their main python modules dependencies. Made .gitignore to help ignore unneeded filed for git in case we need to make env variables in the future.